Modify ArraysTest.MAX_LONG.

+++ src/test/java/edu/marywood/util/ArraysTest.java
@@ -15,7 +15,7 @@ public class ArraysTest {
     static final double MAX_DOUBLE = Double.MAX_VALUE;
     static final float MAX_FLOAT = Float.MAX_VALUE;
     static final int MAX_INT = Integer.MAX_VALUE;
-    static final int MAX_LONG = Long.MAX_VALUE;
+    static final long MAX_LONG = Long.MAX_VALUE;
     static final boolean MIN_BOOLEAN = false;
     static final byte MIN_BYTE = Byte.MIN_VALUE;
     static final char MIN_CHAR = Character.MIN_VALUE;
